Create DLL for openrct2 with win as a wrapper
